chore: Move `server/features/*` routes to `server/*`
