+++ src/v1/server/compliance_tests/routes.rs
@@ -7,36 +7,36 @@ use rocket::{get, put};
 
 /// Get the results of all Compliance Suites executed on your server.
 #[utoipa::path(
-    tag = "Server / Features / Compliance tests",
+    tag = "Server / Compliance tests",
     responses(
         (status = 200, description = "Success", body = String)
     )
 )]
-#[get("/v1/server/features/compliance-tests")]
+#[get("/v1/server/compliance-tests")]
 pub(super) fn get_compliance_tests_results() -> String {
     todo!()
 }
 
 /// Get the results of a Compliance Suite executed on your server.
 #[utoipa::path(
-    tag = "Server / Features / Compliance tests",
+    tag = "Server / Compliance tests",
     responses(
         (status = 200, description = "Success", body = String)
     )
 )]
-#[get("/v1/server/features/compliance-tests/<_xep_id>")]
+#[get("/v1/server/compliance-tests/<_xep_id>")]
 pub(super) fn get_compliance_suite_results(_xep_id: String) -> String {
     todo!()
 }
 
 /// Configure which Compliance Suites to test on your server.
 #[utoipa::path(
-    tag = "Server / Features / Compliance tests",
+    tag = "Server / Compliance tests",
     responses(
         (status = 200, description = "Success", body = String)
     )
 )]
-#[put("/v1/server/features/compliance-tests")]
+#[put("/v1/server/compliance-tests")]
 pub(super) fn set_compliance_test_suites() -> String {
     todo!()
 }
